Optimize chat
