+++ src/components/map/Sidebar/SidebarV2/ChatSidebar.tsx
@@ -1,6 +1,9 @@
-import { Box, Tab, Tabs } from "@mui/material";
+import CloseIcon from "@mui/icons-material/Close";
+import { Box, IconButton, Tab, Tabs } from "@mui/material";
+import { EmojiClickData } from "emoji-picker-react";
 import { getFirestore } from "firebase/firestore";
-import React, { useEffect, useMemo, useState } from "react";
+import dynamic from "next/dynamic";
+import React, { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import { IChannels, IConversation } from "src/chatTypes";
 import { channelsChange, getChannelsSnapshot } from "src/client/firestore/channels.firesrtore";
 import { conversationChange, getConversationsSnapshot } from "src/client/firestore/conversations.firesrtore";
@@ -16,6 +19,11 @@ import { Message } from "../Chat/Room/Message";
 //import { NewsCard } from "../Chat/Room/NewsCard";
 import { SidebarWrapper } from "./SidebarWrapper";
 
+const DynamicMemoEmojiPicker = dynamic(() => import("../Chat/Common/EmojiPicker"), {
+  loading: () => <p>Loading...</p>,
+  ssr: false,
+});
+
 type ChatSidebarProps = {
   open: boolean;
   onClose: () => void;
@@ -30,6 +38,7 @@ type ChatSidebarProps = {
 
 export const ChatSidebar = ({ open, onClose, sidebarWidth, innerHeight, innerWidth, theme }: ChatSidebarProps) => {
   const [value, setValue] = React.useState(0);
+  const [position, setPosition] = useState<{ top: number; left: number }>({ top: 0, left: 0 });
   const [{ user }] = useAuth();
   const handleChange = (event: React.SyntheticEvent, newValue: number) => {
     setValue(newValue);
@@ -39,6 +48,14 @@ export const ChatSidebar = ({ open, onClose, sidebarWidth, innerHeight, innerWid
   const [selectedChannel, setSelectedChannel] = useState("");
   const [channels, setChannels] = useState<IChannels[]>([]);
   const [conversations, setConversations] = useState<IConversation[]>([]);
+  const [showEmojiPicker, setShowEmojiPicker] = useState<boolean>(false);
+  const [reactionsMap, setReactionsMap] = useState<{ [key: string]: string[] }>({});
+  const messageBoxRef = useRef<HTMLDivElement>(null);
+  const messageRef = useRef<{
+    messageId: string | null;
+  }>({
+    messageId: null,
+  });
   const db = getFirestore();
 
   const a11yProps = (index: number) => {
@@ -46,6 +63,73 @@ export const ChatSidebar = ({ open, onClose, sidebarWidth, innerHeight, innerWid
       "aria-controls": `simple-tabpanel-${index}`,
     };
   };
+
+  const toggleEmojiPicker = (event: any, messageId?: string) => {
+    messageRef.current.messageId = messageId || null;
+    const buttonRect = event.target.getBoundingClientRect();
+    const parentDivRect = messageBoxRef?.current?.getBoundingClientRect();
+    let newPosition = {
+      top: buttonRect.bottom + window.scrollY,
+      left: buttonRect.left + window.scrollX,
+    };
+    const left = newPosition.left;
+    const top = newPosition.top;
+    newPosition.left = Math.min(left, (parentDivRect?.right || 0) - window.scrollX - 350);
+    newPosition.top = Math.min(top, (parentDivRect?.bottom || 0) - window.scrollY - 400);
+    if (top >= newPosition.top) {
+      newPosition.top = newPosition.top - 100;
+    }
+    setPosition(newPosition);
+    setShowEmojiPicker(!showEmojiPicker);
+  };
+
+  const handleEmojiClick = useCallback(
+    (emojiObject: EmojiClickData) => {
+      const messageId = messageRef.current.messageId;
+      if (messageId) {
+        toggleReaction(messageId, emojiObject.emoji);
+      } else {
+        // setInputValue(prevValue => prevValue + emojiObject.emoji);
+      }
+      setShowEmojiPicker(false);
+    },
+    [reactionsMap]
+  );
+
+  const addReaction = useCallback(
+    (messageId: string, emoji: string) => {
+      setReactionsMap(prevReactionsMap => ({
+        ...prevReactionsMap,
+        [messageId]: [...(prevReactionsMap[messageId] || []), emoji],
+      }));
+    },
+    [reactionsMap]
+  );
+
+  const removeReaction = useCallback(
+    (messageId: string, emoji: string) => {
+      if (reactionsMap[messageId]) {
+        setReactionsMap(prevReactionsMap => ({
+          ...prevReactionsMap,
+          [messageId]: prevReactionsMap[messageId].filter(reaction => reaction !== emoji),
+        }));
+      }
+    },
+    [reactionsMap]
+  );
+
+  const toggleReaction = useCallback(
+    (messageId: string, emoji: string) => {
+      const messageReactions = reactionsMap[messageId] || [];
+      if (messageReactions.includes(emoji)) {
+        removeReaction(messageId, emoji);
+      } else {
+        addReaction(messageId, emoji);
+      }
+    },
+    [showEmojiPicker, reactionsMap]
+  );
+
   const openRoom = (type: string, channel: any) => {
     setOpenChatRoom(true);
     setRoomType(type);
@@ -65,7 +149,7 @@ export const ChatSidebar = ({ open, onClose, sidebarWidth, innerHeight, innerWid
 
   const contentSignalState = useMemo(() => {
     return { updates: true };
-  }, [openChatRoom, value, roomType]);
+  }, [openChatRoom, value, roomType, position, reactionsMap, showEmojiPicker]);
 
   useEffect(() => {
     if (!user) return;
@@ -103,8 +187,43 @@ export const ChatSidebar = ({ open, onClose, sidebarWidth, innerHeight, innerWid
       sidebarType={"chat"}
       SidebarContent={
         <Box sx={{ borderTop: "solid 1px ", marginTop: openChatRoom ? "9px" : "22px" }}>
+          <Box
+            sx={{
+              display: showEmojiPicker ? "block" : "none",
+              position: "absolute",
+              top: position.top,
+              left: position.left,
+              zIndex: "99",
+            }}
+          >
+            <Box sx={{ display: "flex", justifyContent: "end" }}>
+              <IconButton onClick={() => setShowEmojiPicker(false)}>
+                <CloseIcon />
+              </IconButton>
+            </Box>
+            <DynamicMemoEmojiPicker
+              width="300px"
+              height="400px"
+              onEmojiClick={handleEmojiClick}
+              lazyLoadEmojis={true}
+            />
+          </Box>
           {openChatRoom ? (
-            <Message roomType={roomType} theme={theme} selectedChannel={selectedChannel} user={user} />
+            <Message
+              roomType={roomType}
+              theme={theme}
+              selectedChannel={selectedChannel}
+              user={user}
+              toggleEmojiPicker={toggleEmojiPicker}
+              toggleReaction={toggleReaction}
+              messageBoxRef={messageBoxRef}
+              messageRef={messageRef}
+              position={position}
+              setPosition={setPosition}
+              reactionsMap={reactionsMap}
+              setReactionsMap={setReactionsMap}
+              setShowEmojiPicker={setShowEmojiPicker}
+            />
           ) : (
             <Box>
               <Box
